Fix for adding schedules (#365)
* fix for adding schedules

* Revert "Revert "cost-centers import rewrite (#345)" (#361)"

This reverts commit a9c57aed594602403c03d0554b1a16f84af4a64b.

+++ apps/mappings/imports/schedules.py
@@ -13,12 +13,8 @@ def schedule_or_delete_fyle_import_tasks(configuration: Configuration, mapping_s
     """
     task_to_be_scheduled = None
     # Check if there is a task to be scheduled
-    if mapping_setting_instance:
-        task_to_be_scheduled = MappingSetting.objects.filter(
-            import_to_fyle=True,
-            workspace_id=configuration.workspace_id,
-            source_field=mapping_setting_instance.source_field
-        ).first()
+    if mapping_setting_instance and mapping_setting_instance.import_to_fyle:
+        task_to_be_scheduled = mapping_setting_instance
 
     if task_to_be_scheduled or configuration.import_categories:
         Schedule.objects.update_or_create(
@@ -32,8 +28,15 @@ def schedule_or_delete_fyle_import_tasks(configuration: Configuration, mapping_s
         )
         return
 
-    # If none of the conditions are met, delete the existing schedule
-    Schedule.objects.filter(
-        func='apps.mappings.imports.queues.chain_import_fields_to_fyle',
-        args='{}'.format(configuration.workspace_id)
-    ).delete()
+    import_fields_count = MappingSetting.objects.filter(
+        import_to_fyle=True,
+        workspace_id=configuration.workspace_id, 
+        source_field__in=['CATEGORY', 'PROJECT', 'COST_CENTER']
+    ).count()
+
+    # If the import_fields_count is 0, delete the schedule
+    if import_fields_count == 0:
+        Schedule.objects.filter(
+            func='apps.mappings.imports.queues.chain_import_fields_to_fyle',
+            args='{}'.format(configuration.workspace_id)
+        ).delete()
